was worrying about size_t vs uint32_t stuff.  x64 and x32 builds still work with no changes req.  was wanting to use my fixed point scalar type...  will require a lot of work though.

+++ main/lucid/core/FileReader.cpp
@@ -26,7 +26,7 @@ namespace core {
 		return _file.is_open();
 	}
 
-	void FileReader::read(void *data, int size)
+	void FileReader::read(void *data, size_t size)
 	{
 		LUCID_VALIDATE(is_open(), "attempt to read from an unopened file");
 		_file.read((char *)data, size);
